Fix the logic for determining the consumer threshold
If you have a single consumer and you set your threshold to 0 it would never alert since we checked to see if the count was below the threshold.  Threshold should be the point where you alert not the point above where you alert.  With this fix -C 0 or -W 0 now function as expected instead of silently failing to alert.