Compute queues metrics for damaged queues
We have some queues which have been registered on a node which is not
present on our RabbitMQ cluster anymore.
The queue is still returned when contacting the /queues API endpoint but
most information are missing:

```
  {
    "message_stats"=>
    {
      "publish_details"=>
      {
        "rate"=>0.4
      },
      "publish"=>5816
    },
    "node"=>"rabbit@i-xxxxxxxxxxxxxxxxx.node.consul",
    "arguments"=>{},
    "exclusive"=>false,
    "auto_delete"=>true,
    "durable"=>false,
    "vhost"=>"/",
    "name"=>"i-yyyyyyyyyyyyyyyyy-1.2.0-1522722830",
    "messages"=>0
  }
```

Currently, the `metrics-rabbitmq-queue` facing this kind of queue fails
with:

```
  Check failed to run: undefined method `[]' for nil:NilClass, [
    "bin/metrics-rabbitmq-queue.rb:54:in `block in run'",
    "bin/metrics-rabbitmq-queue.rb:47:in `each'",
    "bin/metrics-rabbitmq-queue.rb:47:in `run'",
    "/home/multani/local/gems/gems/sensu-plugin-1.4.5/lib/sensu-plugin/cli.rb:58:in `block in <class:CLI>'"
  ]
```

and doesn't return any metrics data anymore.

This patch handles such kind of queue and continue the processing of the
rest of the queues as usual.